feat: refactor routes
