+++ static/secure.handlebars
@@ -6,16 +6,16 @@
   </head>
   <body>
     {{ #if edit }}
-    <form action="/data/{{ path }}/{{ token }}" method="POST">
-      <input type="hidden" value="{{ data_type }}" id="data_type" name="data_type">
-      <input type="text" value="{{ data }}" id="data" name="data"><br/>
+    <form action="/data/{{ token }}?css={{ css }}&edit={{ edit }}" method="POST">
+      <input type="hidden" value="{{ path }}" id="path" name="path">
+      {{ data_as_input data }}
       {{ #each key_names }}
-	<label class="checkbox-label"><input type="checkbox" class="checkbox" name="encrypted_by" value="{{ this }}"> {{ this }}</label><br/>
+      <label class="checkbox-label"><input type="checkbox" class="checkbox" name="encryptedby" value="{{ this }}">{{ this }}</label><br/>
       {{ /each }}
       <input type="submit" value="Submit" name="submit" id="submit">
     </form>
     {{ else }}
-    <p>{{ data }}</p>
+    <p>{{ data.value }}</p>
     {{ /if }}
   </body>
 </html>
